feat: migrate to cmake

+++ MikuMikuWorld/Audio/Waveform.h
@@ -6,6 +6,7 @@
 #pragma once
 #include "../Math.h"
 #include "AudioManager.h"
+#include <cstddef>
 #include <stdint.h>
 #include <vector>
 #include <limits>
@@ -46,7 +47,7 @@ namespace Audio
 			const float sampleIndexFraction =
 			    sampleIndexAsFloat - static_cast<int32_t>(sampleIndexAsFloat);
 
-			const size_t sampleIndexLo = std::max(0ull, static_cast<size_t>(sampleIndexAsFloat));
+			const size_t sampleIndexLo = std::max(static_cast<size_t>(0), static_cast<size_t>(sampleIndexAsFloat));
 			const size_t sampleIndexHi = sampleIndexLo + 1;
 
 			const float normalizedSampleLo =
